Move home to Views, remove unused Footer component
