+++ src/views/Home.tsx
@@ -1,6 +1,6 @@
 import React, { CSSProperties } from "react";
 import { Stack, Text, PrimaryButton } from "office-ui-fabric-react";
-import { Links, Images } from "../../common/const";
+import { Links, Images } from "../common/const";
 import { Depths } from "@uifabric/fluent-theme/lib/fluent/FluentDepths";
 import styled from 'styled-components';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
